Fix : 프로필 조회 오류 수정

+++ src/main/java/com/example/rhythme_backend/repository/posts/MakerPostRepository.java
@@ -22,6 +22,6 @@ public interface MakerPostRepository extends JpaRepository<MakerPost,Long> {
     List<MakerPost> findTopByOrderByLikesDesc();
     List<MakerPost> findAllByOrderByLikesDesc();
 
-    Integer countByMember(Member member);
+    Long countByMember(Member member);
 
 }
